feat: Implement new dashboard design and dark theme
This commit introduces a complete redesign of the student and instructor dashboards, creating a modern and responsive layout with a sidebar and top navigation bar.

- **New Layout:** A new layout is established in `layouts/dashboard_layout.html` with supporting CSS in `dashboard_reborn.css`.
- **Student Dashboard:** The student dashboard is updated with a new design that includes a welcome banner, profile completion meter, and quick stats.
- **Instructor Dashboard:** The instructor dashboard is redesigned with a new layout and components.
- **Dark Theme:** A dark theme is implemented using CSS variables and a toggle in the top navigation bar. The styles are defined in `dark_theme.css`.
