--- static/css/discover_feed.css
@@ -1,228 +1,0 @@
-/* Styles for the new Discover Feed */
-
-.discover-container {
-    max-width: 680px;
-    margin: 0 auto;
-    padding: 1rem;
-}
-
-.discover-top-bar {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 0.5rem 1rem;
-    background-color: var(--card-bg);
-    border-bottom: 1px solid #e2e8f0;
-    position: sticky;
-    top: 65px; /* Adjust based on main top bar height */
-    z-index: 999;
-}
-
-.discover-title {
-    font-size: 1.5rem;
-    font-weight: 600;
-}
-
-.stories-row {
-    display: flex;
-    gap: 10px;
-    padding: 1rem 0;
-    overflow-x: auto;
-    border-bottom: 1px solid #e2e8f0;
-    margin-bottom: 1rem;
-}
-
-.story-card {
-    text-align: center;
-    width: 80px;
-    flex-shrink: 0;
-}
-
-.story-card img {
-    width: 60px;
-    height: 60px;
-    border-radius: 50%;
-    border: 2px solid #2D88FF;
-    padding: 2px;
-}
-
-.post-card {
-    background-color: var(--card-bg);
-    border-radius: 8px;
-    margin-bottom: 20px;
-    box-shadow: 0 1px 2px rgba(0,0,0,0.1);
-}
-
-.post-header {
-    display: flex;
-    align-items: center;
-    padding: 1rem;
-}
-
-.post-header img {
-    width: 40px;
-    height: 40px;
-    border-radius: 50%;
-    margin-right: 10px;
-}
-
-.post-author-info .author-name {
-    font-weight: 600;
-}
-
-.post-author-info .timestamp {
-    font-size: 0.8rem;
-    color: #65676b;
-}
-
-.post-text {
-    padding: 0 1rem 1rem 1rem;
-    line-height: 1.5;
-    max-height: 4.5em; /* 3 lines * 1.5 line-height */
-    overflow: hidden;
-    position: relative;
-}
-
-.post-text.expanded {
-    max-height: none;
-}
-
-.see-more-btn {
-    color: #2D88FF;
-    cursor: pointer;
-    font-weight: bold;
-}
-
-.post-media img, .post-media video {
-    width: 100%;
-    max-height: 500px;
-    object-fit: cover;
-}
-
-.image-grid {
-    display: grid;
-    grid-template-columns: repeat(2, 1fr);
-    gap: 2px;
-}
-
-.video-container {
-    position: relative;
-    cursor: pointer;
-}
-
-.play-button-overlay {
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    font-size: 4rem;
-    color: white;
-    background-color: rgba(0,0,0,0.5);
-    border-radius: 50%;
-    width: 80px;
-    height: 80px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    transition: opacity 0.2s;
-}
-
-.video-container:hover .play-button-overlay {
-    opacity: 0.8;
-}
-
-.post-footer {
-    padding: 0.5rem 1rem;
-}
-
-.post-actions {
-    display: flex;
-    justify-content: space-around;
-    padding: 0.5rem 0;
-    border-top: 1px solid #e2e8f0;
-    border-bottom: 1px solid #e2e8f0;
-}
-
-.post-action {
-    display: flex;
-    align-items: center;
-    gap: 5px;
-    color: #65676b;
-    cursor: pointer;
-}
-
-.post-action:hover {
-    color: #2D88FF;
-}
-
-.comment-preview {
-    padding: 1rem 0;
-}
-
-.comment-box {
-    display: flex;
-    align-items: center;
-    gap: 10px;
-    padding: 0.5rem 0;
-}
-
-.comment-box img {
-    width: 32px;
-    height: 32px;
-    border-radius: 50%;
-}
-
-.comment-input {
-    flex: 1;
-    background-color: #f0f2f5;
-    border: none;
-    border-radius: 20px;
-    padding: 8px 12px;
-}
-
-/* Story Viewer Styles */
-.story-viewer-content {
-    background-color: #000;
-    width: 100%;
-    height: 100%;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-#story-content img, #story-content video {
-    max-height: 90vh;
-    max-width: 90vw;
-}
-
-.close-story-viewer {
-    position: absolute;
-    top: 15px;
-    right: 35px;
-    color: #fff;
-    font-size: 40px;
-    font-weight: bold;
-    cursor: pointer;
-    z-index: 1005;
-}
-
-.story-nav {
-    position: absolute;
-    top: 50%;
-    transform: translateY(-50%);
-    background-color: rgba(0,0,0,0.3);
-    color: white;
-    border: none;
-    font-size: 2rem;
-    padding: 0.5rem 1rem;
-    cursor: pointer;
-    z-index: 1004;
-}
-
-.story-nav.prev {
-    left: 10px;
-}
-
-.story-nav.next {
-    right: 10px;
-}
