--- static/css/event-details.css
@@ -1,296 +1,0 @@
-/* Main container for event details */
-.event-details-container {
-    background-color: var(--bg-primary);
-    color: var(--color-primary);
-    padding: 0;
-}
-
-/* Banner image at the top */
-.event-details-banner {
-    width: 100%;
-    height: 350px;
-    overflow: hidden;
-}
-
-.event-details-banner img {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-}
-
-/* Header section below the banner */
-.event-details-header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 2rem;
-    background-color: var(--bg-secondary);
-    border-bottom: 1px solid var(--divider);
-}
-
-.event-details-header h1 {
-    font-size: 2.5rem;
-    font-weight: 700;
-}
-
-.event-details-header .btn {
-    font-weight: 600;
-    padding: 0.75rem 1.5rem;
-}
-
-/* Main body with two columns */
-.event-details-body {
-    display: flex;
-    flex-wrap: wrap;
-    gap: 2rem;
-    padding: 2rem;
-}
-
-.event-main-content {
-    flex: 3; /* Takes up 3 parts of the space */
-    min-width: 300px;
-}
-
-.event-sidebar {
-    flex: 2; /* Takes up 2 parts of the space */
-    min-width: 280px;
-}
-
-/* Card styling for content blocks */
-.event-info-card,
-.event-attendees-card,
-.event-meta-card,
-.event-map-card {
-    background-color: var(--bg-secondary);
-    border-radius: var(--border-radius-lg);
-    padding: 1.5rem;
-    margin-bottom: 2rem;
-    box-shadow: none;
-    border: 1px solid var(--divider);
-}
-
-.event-info-card h3,
-.event-attendees-card h3 {
-    font-size: 1.5rem;
-    font-weight: 600;
-    margin-bottom: 1rem;
-    padding-bottom: 0.5rem;
-    border-bottom: 1px solid var(--divider);
-}
-
-.event-info-card p {
-    line-height: 1.6;
-}
-
-/* Attendees list styling */
-.attendees-list {
-    display: flex;
-    flex-wrap: wrap;
-    gap: 1rem;
-}
-
-.attendee-item {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 0.5rem;
-    width: 80px;
-    text-align: center;
-}
-
-.attendee-item img {
-    width: 60px;
-    height: 60px;
-    border-radius: 50%;
-    object-fit: cover;
-}
-
-.attendee-item span {
-    font-size: 0.875rem;
-    color: var(--color-secondary);
-}
-
-/* Sidebar meta card styling */
-.meta-item {
-    display: flex;
-    align-items: flex-start;
-    gap: 1rem;
-}
-
-.meta-item:not(:last-child) {
-    margin-bottom: 1.5rem;
-}
-
-.meta-item i {
-    font-size: 1.25rem;
-    color: var(--color-accent);
-    margin-top: 5px;
-}
-
-.meta-item div {
-    display: flex;
-    flex-direction: column;
-}
-
-.meta-item strong {
-    font-weight: 600;
-    margin-bottom: 0.25rem;
-}
-
-.meta-item span {
-    color: var(--color-secondary);
-}
-
-.organizer-meta .organizer-img {
-    width: 40px;
-    height: 40px;
-    border-radius: 50%;
-    object-fit: cover;
-}
-
-/* Map card styling */
-#event-map {
-    height: 300px;
-    width: 100%;
-    border-radius: var(--border-radius-lg);
-}
-
-/* Comment Section Styling */
-.event-comments-card {
-    background-color: var(--bg-secondary);
-    border-radius: var(--border-radius-lg);
-    padding: 1.5rem;
-    margin-bottom: 2rem;
-    border: 1px solid var(--divider);
-}
-
-.event-comments-card h3 {
-    font-size: 1.5rem;
-    font-weight: 600;
-    margin-bottom: 1rem;
-    padding-bottom: 0.5rem;
-    border-bottom: 1px solid var(--divider);
-}
-
-.comment-form textarea {
-    width: 100%;
-    padding: 0.75rem;
-    border-radius: var(--border-radius-md);
-    border: 1px solid var(--divider);
-    background-color: var(--bg-tertiary);
-    color: var(--color-primary);
-    resize: vertical;
-    margin-bottom: 0.5rem;
-}
-
-.comment-form {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-end;
-    margin-bottom: 1.5rem;
-}
-
-.comments-list .comment-item {
-    display: flex;
-    gap: 1rem;
-    margin-bottom: 1.5rem;
-}
-
-.comment-author-pic {
-    width: 40px;
-    height: 40px;
-    border-radius: 50%;
-    object-fit: cover;
-}
-
-.comment-content strong {
-    font-weight: 600;
-}
-
-.comment-content .comment-timestamp {
-    font-size: 0.8rem;
-    color: var(--color-secondary);
-    margin-left: 0.5rem;
-}
-
-.comment-content p {
-    margin-top: 0.25rem;
-    line-height: 1.5;
-}
-
-/* "See All" Attendees Modal */
-.modal {
-    display: none;
-    position: fixed;
-    z-index: 1000;
-    left: 0;
-    top: 0;
-    width: 100%;
-    height: 100%;
-    overflow: auto;
-    background-color: rgba(0,0,0,0.6);
-}
-
-.modal-content {
-    background-color: var(--bg-secondary);
-    margin: 10% auto;
-    padding: 2rem;
-    border: 1px solid var(--divider);
-    width: 80%;
-    max-width: 500px;
-    border-radius: var(--border-radius-lg);
-    position: relative;
-}
-
-.close-button {
-    color: var(--color-secondary);
-    position: absolute;
-    top: 1rem;
-    right: 1.5rem;
-    font-size: 28px;
-    font-weight: bold;
-}
-
-.close-button:hover,
-.close-button:focus {
-    color: var(--color-primary);
-    text-decoration: none;
-    cursor: pointer;
-}
-
-.attendees-list-modal {
-    display: flex;
-    flex-direction: column;
-    gap: 1rem;
-    max-height: 60vh;
-    overflow-y: auto;
-    padding-top: 1rem;
-}
-
-.attendee-item-modal {
-    display: flex;
-    align-items: center;
-    gap: 1rem;
-    padding: 0.5rem;
-    border-radius: var(--border-radius-md);
-    text-decoration: none;
-    color: var(--color-primary);
-    transition: background-color 0.2s ease;
-}
-
-.attendee-item-modal:hover {
-    background-color: var(--bg-tertiary);
-}
-
-.attendee-item-modal img {
-    width: 50px;
-    height: 50px;
-    border-radius: 50%;
-    object-fit: cover;
-}
-
-.attendees-header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-}
