--- static/css/events.css
@@ -1,296 +1,0 @@
-/* New three-column layout for the main events page */
-.events-page-container {
-    display: flex;
-    gap: 1.5rem;
-    padding: 1rem;
-}
-
-.events-sidebar {
-    flex: 1;
-    min-width: 220px;
-}
-
-.events-main-content {
-    flex: 3;
-    min-width: 0; /* Prevents oversized content from breaking flexbox */
-}
-
-.right-sidebar {
-    flex: 1.5;
-    min-width: 280px;
-}
-
-/* Styling for the sidebar sections */
-.sidebar-section {
-    background-color: var(--bg-secondary);
-    padding: 1rem;
-    border-radius: var(--border-radius-lg);
-    margin-bottom: 1.5rem;
-    border: 1px solid var(--divider);
-}
-
-.sidebar-section h4 {
-    font-size: 1.1rem;
-    font-weight: 600;
-    margin-bottom: 1rem;
-    padding-bottom: 0.5rem;
-    border-bottom: 1px solid var(--divider);
-}
-
-.sidebar-list {
-    list-style: none;
-    padding: 0;
-    margin: 0;
-}
-
-.sidebar-list li a {
-    display: block;
-    padding: 0.5rem 0;
-    color: var(--color-secondary);
-    text-decoration: none;
-    transition: color 0.2s ease;
-}
-
-.sidebar-list li a:hover {
-    color: var(--color-accent);
-}
-
-#calendar-placeholder {
-    height: 250px;
-    background-color: var(--bg-tertiary);
-    border-radius: var(--border-radius-md);
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    color: var(--color-secondary);
-}
-
-/* Header section with title and create button */
-.events-header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    margin-bottom: 2rem;
-    padding-bottom: 1rem;
-    border-bottom: 1px solid var(--divider);
-}
-
-.events-header h1 {
-    font-size: 2rem;
-    font-weight: 700;
-}
-
-#create-event-btn {
-    display: flex;
-    align-items: center;
-    gap: 0.5rem;
-    font-weight: 600;
-}
-
-/* Search and filter controls */
-.events-controls {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    margin-bottom: 2rem;
-    gap: 1rem;
-}
-
-.search-bar {
-    position: relative;
-    flex-grow: 1;
-}
-
-.search-bar input {
-    width: 100%;
-    padding: 0.75rem 1rem 0.75rem 2.5rem; /* Add padding for icon */
-    border-radius: var(--border-radius-lg);
-    border: 1px solid var(--divider);
-    background-color: var(--bg-secondary);
-    color: var(--color-primary);
-}
-
-.search-bar .fa-search {
-    position: absolute;
-    left: 1rem;
-    top: 50%;
-    transform: translateY(-50%);
-    color: var(--color-secondary);
-}
-
-.filters select {
-    padding: 0.75rem 1rem;
-    border-radius: var(--border-radius-lg);
-    border: 1px solid var(--divider);
-    background-color: var(--bg-secondary);
-    color: var(--color-primary);
-}
-
-/* Grid for event cards */
-.events-grid {
-    display: grid;
-    grid-template-columns: repeat(auto-fill, minmax(300px, 1fr));
-    gap: 1.5rem;
-}
-
-/* Individual event card styling */
-.event-card {
-    background-color: var(--bg-secondary);
-    border-radius: var(--border-radius-lg);
-    overflow: hidden;
-    box-shadow: none;
-    border: 1px solid var(--divider);
-    display: flex;
-    flex-direction: column;
-    transition: transform 0.2s ease, box-shadow 0.2s ease;
-}
-
-.event-card:hover {
-    transform: translateY(-5px);
-    border-color: var(--color-accent);
-}
-
-.event-card-banner {
-    width: 100%;
-    height: 180px;
-    overflow: hidden;
-    position: relative;
-}
-
-.live-badge {
-    position: absolute;
-    top: 10px;
-    left: 10px;
-    background-color: #e74c3c;
-    color: white;
-    padding: 5px 10px;
-    border-radius: 5px;
-    font-size: 0.8rem;
-    font-weight: 700;
-    animation: pulse 1.5s infinite;
-}
-
-@keyframes pulse {
-    0% {
-        box-shadow: 0 0 0 0 rgba(231, 76, 60, 0.7);
-    }
-    70% {
-        box-shadow: 0 0 0 10px rgba(231, 76, 60, 0);
-    }
-    100% {
-        box-shadow: 0 0 0 0 rgba(231, 76, 60, 0);
-    }
-}
-
-.event-card-banner img {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-}
-
-.event-card-content {
-    padding: 1.25rem;
-    flex-grow: 1;
-}
-
-.event-title {
-    font-size: 1.25rem;
-    font-weight: 600;
-    margin-bottom: 0.75rem;
-}
-
-.event-date, .event-location {
-    display: flex;
-    align-items: center;
-    gap: 0.5rem;
-    color: var(--color-secondary);
-    margin-bottom: 0.5rem;
-}
-
-.event-organizer {
-    display: flex;
-    align-items: center;
-    gap: 0.5rem;
-    margin-top: 1rem;
-}
-
-.event-organizer img {
-    width: 30px;
-    height: 30px;
-    border-radius: 50%;
-    object-fit: cover;
-}
-
-.event-organizer span {
-    font-weight: 500;
-}
-
-.event-card-footer {
-    padding: 1.25rem;
-    border-top: 1px solid var(--divider);
-    display: flex;
-    justify-content: flex-end;
-}
-
-.no-events {
-    grid-column: 1 / -1; /* Span full width */
-    text-align: center;
-    padding: 3rem;
-    background-color: var(--bg-secondary);
-    border-radius: var(--border-radius-lg);
-}
-
-/* Floating Action Button for mobile */
-.fab {
-    display: none; /* Hidden by default */
-    position: fixed;
-    bottom: 20px;
-    right: 20px;
-    width: 60px;
-    height: 60px;
-    background-color: var(--color-accent);
-    color: white;
-    border-radius: 50%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    font-size: 24px;
-    box-shadow: 0 4px 8px rgba(0,0,0,0.2);
-    z-index: 1000;
-}
-
-/* Responsive Design */
-@media (max-width: 1024px) {
-    .events-page-container {
-        flex-direction: column;
-    }
-    .right-sidebar {
-        display: none; /* Hide calendar on tablets */
-    }
-    .events-main-content {
-        order: 1;
-    }
-    .left-sidebar {
-        order: 2;
-    }
-}
-
-@media (max-width: 768px) {
-    .left-sidebar {
-        display: none; /* Hide left sidebar on mobile */
-    }
-    .events-header {
-        flex-direction: column;
-        align-items: flex-start;
-        gap: 1rem;
-    }
-    .events-header .btn {
-        display: none; /* Hide create button, use FAB instead */
-    }
-    .fab {
-        display: flex; /* Show FAB on mobile */
-    }
-    .events-grid {
-        grid-template-columns: 1fr; /* Single column on mobile */
-    }
-}
