--- static/css/resources.css
@@ -1,75 +1,0 @@
-/* Styles for the Resources (Marketplace) Page */
-
-.resources-grid {
-    display: grid;
-    grid-template-columns: repeat(auto-fill, minmax(250px, 1fr));
-    gap: 1.5rem;
-}
-
-.resource-card {
-    background-color: var(--card-bg);
-    border-radius: 8px;
-    overflow: hidden;
-    box-shadow: 0 1px 3px rgba(0,0,0,0.12), 0 1px 2px rgba(0,0,0,0.24);
-    transition: all 0.3s cubic-bezier(.25,.8,.25,1);
-}
-
-.resource-card:hover {
-    box-shadow: 0 14px 28px rgba(0,0,0,0.25), 0 10px 10px rgba(0,0,0,0.22);
-}
-
-.resource-thumbnail {
-    width: 100%;
-    height: 150px;
-    background-color: #e9ecef;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    font-size: 3rem;
-    color: #6c757d;
-}
-
-.resource-thumbnail img {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-}
-
-.resource-info {
-    padding: 1rem;
-}
-
-.resource-title {
-    font-size: 1.1rem;
-    font-weight: 600;
-    margin-bottom: 0.5rem;
-}
-
-.resource-description {
-    font-size: 0.9rem;
-    color: #6c757d;
-    height: 3.6em; /* 3 lines */
-    overflow: hidden;
-}
-
-.resource-price {
-    font-size: 1.2rem;
-    font-weight: bold;
-    color: #2D88FF;
-    margin-top: 1rem;
-}
-
-.resource-seller {
-    display: flex;
-    align-items: center;
-    margin-top: 1rem;
-    padding-top: 1rem;
-    border-top: 1px solid #e9ecef;
-}
-
-.resource-seller img {
-    width: 30px;
-    height: 30px;
-    border-radius: 50%;
-    margin-right: 10px;
-}
