--- static/library/18f7f2aa2ac001b5.pdf
@@ -1,1 +1,0 @@
-this is a test file
