--- static/library/2cfe048d0c83bf32.pdf
@@ -1,1 +1,0 @@
-this is a test file
