--- static/library/5f12827b4bb867f5.pdf
@@ -1,1 +1,0 @@
-this is a test file
