--- static/library/6112673f3a1a2e35.pdf
@@ -1,1 +1,0 @@
-this is a test file
