--- static/library/6fa57f0362a40f3f.pdf
@@ -1,1 +1,0 @@
-this is a test file
