--- static/library/9d5af44ed69c2dfb.pdf
@@ -1,1 +1,0 @@
-this is a test file
