--- static/library/c78b15508b51e5e6.pdf
@@ -1,1 +1,0 @@
-this is a test file
