--- static/payment_proofs/2f04e7558166c16c.pdf
@@ -1,1 +1,0 @@
-proof
