--- static/payment_proofs/33a46c17cc09a092.png
@@ -1,1 +1,0 @@
-payment proof
