--- static/payment_proofs/51802bc9204b53c4.png
@@ -1,1 +1,0 @@
-payment proof
