--- static/payment_proofs/9f3d28233722f446.pdf
@@ -1,1 +1,0 @@
-proof
