--- static/payment_proofs/a2262050c46f257d.png
@@ -1,1 +1,0 @@
-payment proof
