--- static/payment_proofs/c9d5b2ee2c28b0b9.png
@@ -1,1 +1,0 @@
-payment proof
