--- static/payment_proofs/cbbeb6bcd6b01cea.pdf
@@ -1,1 +1,0 @@
-proof
