--- static/payment_proofs/fddf87ff538555f2.pdf
@@ -1,1 +1,0 @@
-proof
