--- static/uploads/images/4861a9130e461970.jpg
@@ -1,1 +1,0 @@
-fake image data
