--- static/uploads/images/cb2f9d9a3df3d08e.jpg
@@ -1,1 +1,0 @@
-fake image data
