--- static/uploads/images/d6dce7f88ba34410.jpg
@@ -1,1 +1,0 @@
-fake image data
