--- static/uploads/images/e1e3995340af9eaa.jpg
@@ -1,1 +1,0 @@
-fake image data
