--- templates/events/details.html
@@ -1,143 +1,0 @@
-{% extends "layouts/dashboard_layout.html" %}
-
-{% block title %}{{ event.title }}{% endblock %}
-
-{% block head %}
-  {{ super() }}
-  <link rel="stylesheet" href="{{ url_for('static', filename='css/event-details.css') }}">
-  {# Add leaflet CSS for map #}
-  <link rel="stylesheet" href="https://unpkg.com/leaflet@1.7.1/dist/leaflet.css" />
-{% endblock %}
-
-{% block content %}
-<div class="event-details-container">
-    <div class="event-details-banner">
-        <img src="{{ url_for('static', filename=event.image or 'images/course_placeholder.jpg') }}" alt="{{ event.title }} Banner">
-    </div>
-
-    <div class="event-details-header">
-        <h1>{{ event.title }}</h1>
-        <form action="{{ url_for('main.event_rsvp', event_id=event.id) }}" method="POST">
-            {% if user_rsvp %}
-                <button type="submit" class="btn btn-danger"><i class="fas fa-times-circle"></i> Cancel RSVP</button>
-            {% else %}
-                <button type="submit" class="btn btn-success"><i class="fas fa-check-circle"></i> RSVP Now</button>
-            {% endif %}
-        </form>
-    </div>
-
-    <div class="event-details-body">
-        <div class="event-main-content">
-            <div class="event-info-card">
-                <h3>About this Event</h3>
-                <p>{{ event.description | safe }}</p>
-            </div>
-
-            <div class="event-attendees-card">
-                <div class="attendees-header">
-                    <h3>Attendees ({{ event.rsvps.count() }})</h3>
-                    {% if event.rsvps.count() > 10 %}
-                        <button class="btn btn-sm btn-secondary" id="see-all-attendees-btn">See All</button>
-                    {% endif %}
-                </div>
-                <div class="attendees-list">
-                    {% for rsvp in event.rsvps.limit(10) %}
-                        <div class="attendee-item">
-                            <img src="{{ url_for('static', filename='profile_pics/' + (rsvp.user.profile_pic or 'default.jpg')) }}" alt="{{ rsvp.user.name }}">
-                            <span>{{ rsvp.user.name }}</span>
-                        </div>
-                    {% else %}
-                        <p>Be the first to RSVP!</p>
-                    {% endfor %}
-                </div>
-            </div>
-
-            <div class="event-comments-card">
-                <h3>Comments ({{ comments|length }})</h3>
-                <div class="comment-form">
-                    <form action="{{ url_for('main.post_event_comment', event_id=event.id) }}" method="POST">
-                        <textarea name="comment_text" placeholder="Add a public comment..." required></textarea>
-                        <button type="submit" class="btn btn-primary">Comment</button>
-                    </form>
-                </div>
-                <div class="comments-list">
-                    {% for comment in comments %}
-                        <div class="comment-item">
-                            <img src="{{ url_for('static', filename='profile_pics/' + (comment.author.profile_pic or 'default.jpg')) }}" alt="{{ comment.author.name }}" class="comment-author-pic">
-                            <div class="comment-content">
-                                <strong>{{ comment.author.name }}</strong>
-                                <span class="comment-timestamp">{{ comment.timestamp.strftime('%b %d, %Y') }}</span>
-                                <p>{{ comment.content }}</p>
-                            </div>
-                        </div>
-                    {% else %}
-                        <p>No comments yet. Be the first to comment!</p>
-                    {% endfor %}
-                </div>
-            </div>
-        </div>
-
-        <div class="event-sidebar">
-            <div class="event-meta-card">
-                <div class="meta-item">
-                    <i class="fas fa-calendar-alt"></i>
-                    <div>
-                        <strong>Date & Time</strong>
-                        <span>{{ event.date.strftime('%A, %B %d, %Y') }}</span>
-                        <span>{{ event.date.strftime('%I:%M %p') }}</span>
-                    </div>
-                </div>
-                <div class="meta-item">
-                    <i class="fas fa-map-marker-alt"></i>
-                    <div>
-                        <strong>Location</strong>
-                        <span>{{ event.location }}</span>
-                    </div>
-                </div>
-                 <div class="meta-item organizer-meta">
-                    <img src="{{ url_for('static', filename='profile_pics/' + (event.organizer.profile_pic or 'default.jpg')) }}" alt="{{ event.organizer.name }}" class="organizer-img">
-                    <div>
-                        <strong>Organizer</strong>
-                        <span>{{ event.organizer.name }}</span>
-                    </div>
-                </div>
-            </div>
-            <div class="event-map-card">
-                <div id="event-map"></div>
-            </div>
-        </div>
-    </div>
-</div>
-{% endblock %}
-
-{% block scripts %}
-    {{ super() }}
-    {# Add leaflet JS for map #}
-    <script src="https://unpkg.com/leaflet@1.7.1/dist/leaflet.js"></script>
-    <script src="{{ url_for('static', filename='js/event_details.js') }}"></script>
-    <script>
-        // Basic map initialization - in a real app, you'd geocode the location string
-        document.addEventListener('DOMContentLoaded', function () {
-            var map = L.map('event-map').setView([51.505, -0.09], 13); // Default view
-            L.tileLayer('https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png', {
-                attribution: '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors'
-            }).addTo(map);
-
-            // In a real app, you'd use a geocoding service to get lat/lng from event.location
-            // For now, we just add a marker at the default location
-            L.marker([51.505, -0.09]).addTo(map)
-                .bindPopup('{{ event.location | e }}')
-                .openPopup();
-        });
-    </script>
-{% endblock %}
-
-<div id="attendees-modal" class="modal">
-    <div class="modal-content">
-        <span class="close-button">&times;</span>
-        <h2>All Attendees</h2>
-        <div id="modal-attendees-list" class="attendees-list-modal">
-            <!-- Attendees will be populated by JavaScript -->
-        </div>
-    </div>
-</div>
