+++ templates/layouts/dashboard_layout.html
@@ -43,21 +43,7 @@
                              <li><a href="{{ url_for('admin.dashboard') }}"><i class="fas fa-tachometer-alt"></i><span>Dashboard</span></a></li>
                         {% endif %}
 
-                        <li><a href="{{ url_for('feed.discover') }}"><i class="fas fa-search"></i><span>Discover</span></a></li>
-                        <li><a href="{{ url_for('feed.home') }}"><i class="fas fa-rss"></i><span>Feed</span></a></li>
-                        <li><a href="{{ url_for('main.events') }}"><i class="fas fa-calendar-alt"></i><span>Events</span></a></li>
-                        <li><a href="{{ url_for('main.chat_list') }}"><i class="fas fa-comments"></i><span>Chat</span></a></li>
-
-                        <li class="nav-item-dropdown">
-                            <a href="#" class="dropdown-toggle"><i class="fas fa-book-reader"></i><span>Library</span></a>
-                            <ul class="dropdown-menu">
-                                <li><a href="{{ url_for('main.library') }}">Explore Library</a></li>
-                                {% if current_user.role == 'instructor' %}
-                                <li><a href="#" class="upload-library-trigger">Upload to Library</a></li>
-                                {% endif %}
-                            </ul>
-                        </li>
-                        <li><a href="{{ url_for('more.resources') }}"><i class="fas fa-cogs"></i><span>Resources</span></a></li>
+                        <li><a href="#"><i class="fas fa-comments"></i><span>Chat</span></a></li>
                     {% endif %}
                 </ul>
             </nav>
