--- templates/more/resource_detail.html
@@ -1,40 +1,0 @@
-{% extends "layouts/dashboard_layout.html" %}
-
-{% block title %}{{ material.title }}{% endblock %}
-
-{% block dashboard_content %}
-    <div class="resource-detail-container">
-        <h1>{{ material.title }}</h1>
-
-        <div class="resource-seller" style="margin-bottom: 2rem;">
-            <img src="{{ url_for('static', filename='profile_pics/' + material.uploader.profile_pic) }}" alt="{{ material.uploader.name }}'s profile picture" style="width: 50px; height: 50px; border-radius: 50%; margin-right: 1rem;">
-            <div>
-                <strong>Seller:</strong> {{ material.uploader.name }}<br>
-                <a href="#" class="btn-secondary">Chat with seller</a>
-            </div>
-        </div>
-
-        <div class="card">
-            <div class="card-header">Description</div>
-            <p>{{ material.description }}</p>
-        </div>
-
-        <div class="card" style="margin-top: 2rem;">
-            <div class="card-header">Purchase Information</div>
-            <div class="resource-price" style="font-size: 1.5rem; margin-bottom: 1rem;">
-                {% if material.price_naira > 0 %}
-                    Price: ₦{{ material.price_naira }}
-                {% else %}
-                    Price: Free
-                {% endif %}
-            </div>
-            <a href="#" class="btn-primary">
-                {% if material.price_naira > 0 %}
-                    Buy Now
-                {% else %}
-                    Download
-                {% endif %}
-            </a>
-        </div>
-    </div>
-{% endblock %}
