--- templates/placeholder.html
@@ -1,10 +1,0 @@
-{% extends "layouts/dashboard_layout.html" %}
-
-{% block title %}Coming Soon{% endblock %}
-
-{% block content %}
-<div class="card">
-    <div class="card-header">Coming Soon!</div>
-    <p>This feature is currently under development. Please check back later.</p>
-</div>
-{% endblock %}
